+++ docs/changelog.rst
@@ -3,6 +3,8 @@
 
 Changelog
 =========
+- :release:`9.7.0 <10th June 2023>`
+- :feature:`179` Add paste service utility to upload text to our paste service.
 - :feature:`176` Migrate repo to use ruff for linting